+++ apps/desktop/src/main/services/__tests__/MainProcessServices.test.ts
@@ -1,7 +1,7 @@
-import { MainProcessServices } from "../MainProcessServices";
-import { NodeFileSystemBridge } from "../NodeFileSystemBridge";
 import { NodeCryptoUtils } from "../../utils/NodeCryptoUtils";
 import { NodeDeviceInfo } from "../../utils/NodeDeviceInfo";
+import { MainProcessServices } from "../MainProcessServices";
+import { NodeFileSystemBridge } from "../NodeFileSystemBridge";
 
 // Mock electron app module
 jest.mock("electron", () => ({
@@ -107,14 +107,9 @@ describe("MainProcessServices", () => {
       );
     });
 
-    it("should initialize MigrationService", () => {
-      expect(services.migrationService).toBeDefined();
-      expect(MockedMigrationService).toHaveBeenCalledWith(
-        services.databaseBridge,
-        services.fileSystemBridge,
-        expect.any(Object), // pathUtils
-        expect.any(String), // migrations path
-      );
+    it("should initialize MainDatabaseService", () => {
+      expect(services.databaseService).toBeDefined();
+      expect(services.databaseService).toHaveProperty("runDatabaseMigrations");
     });
   });
 
@@ -158,7 +153,7 @@ describe("MainProcessServices", () => {
     }
 
     it("should create a database service with the configured database bridge", () => {
-      const userRepository = services.createDatabaseService(
+      const userRepository = services.createDatabaseBridge(
         (db) => new MockUserRepository(db),
       );
 
@@ -170,7 +165,7 @@ describe("MainProcessServices", () => {
     it("should pass the database bridge to the service factory", () => {
       const mockFactory = jest.fn((db) => new MockUserRepository(db));
 
-      const service = services.createDatabaseService(mockFactory);
+      const service = services.createDatabaseBridge(mockFactory);
 
       expect(mockFactory).toHaveBeenCalledWith(services.databaseBridge);
       expect(mockFactory).toHaveBeenCalledTimes(1);
@@ -178,10 +173,10 @@ describe("MainProcessServices", () => {
     });
 
     it("should create different service instances for different factory calls", () => {
-      const service1 = services.createDatabaseService(
+      const service1 = services.createDatabaseBridge(
         (db) => new MockUserRepository(db),
       );
-      const service2 = services.createDatabaseService(
+      const service2 = services.createDatabaseBridge(
         (db) => new MockUserRepository(db),
       );
 
@@ -190,12 +185,12 @@ describe("MainProcessServices", () => {
     });
 
     it("should support different service types with generics", () => {
-      const userRepo = services.createDatabaseService<MockUserRepository>(
+      const userRepo = services.createDatabaseBridge<MockUserRepository>(
         (db) => new MockUserRepository(db),
       );
 
       const conversationService =
-        services.createDatabaseService<MockConversationService>(
+        services.createDatabaseBridge<MockConversationService>(
           (db) => new MockConversationService(db, services.logger),
         );
 
@@ -211,12 +206,12 @@ describe("MainProcessServices", () => {
       };
 
       expect(() => {
-        services.createDatabaseService(errorFactory);
+        services.createDatabaseBridge(errorFactory);
       }).toThrow("Service creation failed");
     });
 
     it("should maintain type safety for created services", () => {
-      const userRepository = services.createDatabaseService(
+      const userRepository = services.createDatabaseBridge(
         (db) => new MockUserRepository(db),
       );
 
@@ -227,7 +222,7 @@ describe("MainProcessServices", () => {
 
     it("should support complex service factory patterns", () => {
       // Test with additional dependencies passed to service
-      const complexService = services.createDatabaseService((db) => {
+      const complexService = services.createDatabaseBridge((db) => {
         const repository = new MockUserRepository(db);
         const service = new MockConversationService(db, services.logger);
 
@@ -393,607 +388,36 @@ describe("MainProcessServices", () => {
     });
   });
 
-  describe("performDatabaseHealthCheck", () => {
-    let mockDatabaseBridge: any;
-
-    beforeEach(() => {
-      mockDatabaseBridge = services.databaseBridge;
-    });
-
-    it("should return healthy when database is connected and query succeeds", async () => {
-      mockDatabaseBridge.isConnected.mockReturnValue(true);
-      mockDatabaseBridge.query.mockResolvedValue([{ test: 1 }]);
-
-      const result = await services.performDatabaseHealthCheck();
-
-      expect(result.isHealthy).toBe(true);
-      expect(result.issues).toEqual([]);
-      expect(mockDatabaseBridge.isConnected).toHaveBeenCalled();
-      expect(mockDatabaseBridge.query).toHaveBeenCalledWith(
-        "SELECT 1 as test",
-        [],
-      );
-    });
-
-    it("should return unhealthy when database is not connected", async () => {
-      mockDatabaseBridge.isConnected.mockReturnValue(false);
-
-      const result = await services.performDatabaseHealthCheck();
-
-      expect(result.isHealthy).toBe(false);
-      expect(result.issues).toContain("Database connection not established");
-      expect(mockDatabaseBridge.query).not.toHaveBeenCalled();
-    });
-
-    it("should return unhealthy when query fails", async () => {
-      mockDatabaseBridge.isConnected.mockReturnValue(true);
-      mockDatabaseBridge.query.mockRejectedValue(new Error("Query failed"));
-
-      const result = await services.performDatabaseHealthCheck();
-
-      expect(result.isHealthy).toBe(false);
-      expect(result.issues).toContain(
-        "Database health check failed: Query failed",
-      );
-    });
-
-    it("should return unhealthy when query returns no results", async () => {
-      mockDatabaseBridge.isConnected.mockReturnValue(true);
-      mockDatabaseBridge.query.mockResolvedValue([]);
-
-      const result = await services.performDatabaseHealthCheck();
-
-      expect(result.isHealthy).toBe(false);
-      expect(result.issues).toContain(
-        "Database health check failed: Database connectivity test failed",
-      );
-    });
-
-    it("should handle non-Error exceptions", async () => {
-      mockDatabaseBridge.isConnected.mockReturnValue(true);
-      mockDatabaseBridge.query.mockRejectedValue("String error");
-
-      const result = await services.performDatabaseHealthCheck();
-
-      expect(result.isHealthy).toBe(false);
-      expect(result.issues).toContain(
-        "Database health check failed: Unknown database error",
-      );
-    });
-  });
-
   describe("runDatabaseMigrations", () => {
-    let mockMigrationService: any;
+    let mockRunDatabaseMigrations: jest.SpyInstance;
 
     beforeEach(() => {
-      mockMigrationService = services.migrationService;
+      // Mock the database service method
+      mockRunDatabaseMigrations = jest
+        .spyOn(services.databaseService, "runDatabaseMigrations")
+        .mockResolvedValue(undefined);
     });
 
-    it("should run migrations successfully", async () => {
-      const mockResult = {
-        success: true,
-        migrationsRun: 2,
-        currentVersion: 2,
-      };
-      mockMigrationService.runMigrations.mockResolvedValue(mockResult);
+    afterEach(() => {
+      mockRunDatabaseMigrations.mockRestore();
+    });
+
+    it("should delegate to database service", async () => {
+      mockRunDatabaseMigrations.mockResolvedValue(undefined);
 
       await services.runDatabaseMigrations();
 
-      expect(mockMigrationService.runMigrations).toHaveBeenCalled();
+      expect(mockRunDatabaseMigrations).toHaveBeenCalled();
     });
 
-    it("should handle migration failures with error details", async () => {
-      const mockResult = {
-        success: false,
-        migrationsRun: 1,
-        currentVersion: 1,
-        errors: [
-          {
-            order: 2,
-            filename: "002_add_users_table.sql",
-            error: "Table already exists",
-          },
-        ],
-      };
-      mockMigrationService.runMigrations.mockResolvedValue(mockResult);
+    it("should propagate database service errors", async () => {
+      const mockError = new Error("Database service failed");
+      mockRunDatabaseMigrations.mockRejectedValue(mockError);
 
       await expect(services.runDatabaseMigrations()).rejects.toThrow(
-        "Database migrations failed: 002_add_users_table.sql: Table already exists",
+        "Database service failed",
       );
-      expect(mockMigrationService.runMigrations).toHaveBeenCalled();
-    });
-
-    it("should handle migration service exceptions", async () => {
-      const mockError = new Error("Migration service failed");
-      mockMigrationService.runMigrations.mockRejectedValue(mockError);
-
-      await expect(services.runDatabaseMigrations()).rejects.toThrow(
-        "Migration execution failed: Migration service failed",
-      );
-      expect(mockMigrationService.runMigrations).toHaveBeenCalled();
-    });
-
-    it("should handle non-Error exceptions", async () => {
-      mockMigrationService.runMigrations.mockRejectedValue("String error");
-
-      await expect(services.runDatabaseMigrations()).rejects.toThrow(
-        "Migration execution failed: String error",
-      );
-      expect(mockMigrationService.runMigrations).toHaveBeenCalled();
-    });
-  });
-
-  describe("migration path methods", () => {
-    describe("getSourceMigrationsPath", () => {
-      it("should return app resources path for packaged app", () => {
-        // Mock app.isPackaged as true
-        const { app } = require("electron");
-        app.isPackaged = true;
-
-        // Need to recreate services to pick up the mocked isPackaged value
-        const packagedServices = new MainProcessServices();
-
-        // Access the private method via bracket notation for testing
-        const sourcePath = (packagedServices as any).getSourceMigrationsPath();
-
-        expect(sourcePath).toBe("/mock/resources/migrations");
-        expect(app.isPackaged).toBe(true);
-      });
-
-      it("should return project root path for development app", () => {
-        const { app } = require("electron");
-        app.isPackaged = false;
-        app.getAppPath = jest.fn(
-          () => "/dev/fishbowl/apps/desktop/dist-electron/electron",
-        );
-
-        const devServices = new MainProcessServices();
-        const sourcePath = (devServices as any).getSourceMigrationsPath();
-
-        expect(sourcePath).toBe("/dev/fishbowl/migrations");
-        expect(app.getAppPath).toHaveBeenCalled();
-      });
-
-      it("should log appropriate debug information", () => {
-        const { app } = require("electron");
-        app.isPackaged = true;
-
-        const testServices = new MainProcessServices();
-        const mockLogger = {
-          debug: jest.fn(),
-          info: jest.fn(),
-          warn: jest.fn(),
-          error: jest.fn(),
-        };
-        (testServices as any).logger = mockLogger;
-
-        (testServices as any).getSourceMigrationsPath();
-
-        expect(mockLogger.debug).toHaveBeenCalledWith(
-          "Using packaged migrations source path",
-          { path: "/mock/resources/migrations" },
-        );
-      });
-    });
-
-    describe("getMigrationsPath", () => {
-      it("should always return userData/migrations path", () => {
-        const { app } = require("electron");
-        app.getPath = jest.fn((type: string) => {
-          if (type === "userData") return "/user/data";
-          return "/mock/path";
-        });
-
-        const testServices = new MainProcessServices();
-        const migrationsPath = (testServices as any).getMigrationsPath();
-
-        expect(migrationsPath).toBe("/user/data/migrations");
-        expect(app.getPath).toHaveBeenCalledWith("userData");
-      });
-
-      it("should log debug information", () => {
-        const { app } = require("electron");
-        app.getPath = jest.fn((type: string) => {
-          if (type === "userData") return "/mock/userdata";
-          return "/mock/path";
-        });
-
-        const testServices = new MainProcessServices();
-        const mockLogger = {
-          debug: jest.fn(),
-          info: jest.fn(),
-          warn: jest.fn(),
-          error: jest.fn(),
-        };
-        (testServices as any).logger = mockLogger;
-
-        (testServices as any).getMigrationsPath();
-
-        expect(mockLogger.debug).toHaveBeenCalledWith(
-          "Using userData migrations path",
-          { path: "/mock/userdata/migrations" },
-        );
-      });
-    });
-
-    describe("validateMigrationPaths", () => {
-      let testServices: MainProcessServices;
-      let mockFileSystemBridge: any;
-      let mockLogger: any;
-
-      beforeEach(() => {
-        testServices = new MainProcessServices();
-        mockLogger = {
-          debug: jest.fn(),
-          info: jest.fn(),
-          warn: jest.fn(),
-          error: jest.fn(),
-        };
-        mockFileSystemBridge = {
-          getDirectoryStats: jest.fn(),
-          readdir: jest.fn(),
-        };
-        (testServices as any).logger = mockLogger;
-        (testServices as any).fileSystemBridge = mockFileSystemBridge;
-      });
-
-      it("should validate existing source directory with SQL files", async () => {
-        mockFileSystemBridge.getDirectoryStats.mockResolvedValue({
-          exists: true,
-          isDirectory: true,
-        });
-        mockFileSystemBridge.readdir.mockResolvedValue([
-          "001_initial.sql",
-          "002_add_users.sql",
-          "readme.txt",
-        ]);
-
-        await expect(
-          (testServices as any).validateMigrationPaths(
-            "/source/migrations",
-            "/mock/userdata/migrations",
-          ),
-        ).resolves.toBeUndefined();
-
-        expect(mockFileSystemBridge.getDirectoryStats).toHaveBeenCalledWith(
-          "/source/migrations",
-        );
-        expect(mockFileSystemBridge.readdir).toHaveBeenCalledWith(
-          "/source/migrations",
-        );
-        expect(mockLogger.debug).toHaveBeenCalledWith(
-          "Migration paths validated successfully",
-          expect.objectContaining({
-            sourcePath: "/source/migrations",
-            sqlFilesFound: 2,
-          }),
-        );
-      });
-
-      it("should throw error if source directory does not exist", async () => {
-        mockFileSystemBridge.getDirectoryStats.mockResolvedValue({
-          exists: false,
-          isDirectory: false,
-        });
-
-        await expect(
-          (testServices as any).validateMigrationPaths(
-            "/nonexistent/migrations",
-            "/mock/userdata/migrations",
-          ),
-        ).rejects.toThrow("Source migrations directory not found");
-
-        expect(mockLogger.warn).toHaveBeenCalledWith(
-          "Source migrations path does not exist or is not a directory",
-          expect.objectContaining({
-            sourcePath: "/nonexistent/migrations",
-            exists: false,
-            isDirectory: false,
-          }),
-        );
-      });
-
-      it("should throw error if destination path is outside userData", async () => {
-        mockFileSystemBridge.getDirectoryStats.mockResolvedValue({
-          exists: true,
-          isDirectory: true,
-        });
-
-        await expect(
-          (testServices as any).validateMigrationPaths(
-            "/source/migrations",
-            "/dangerous/path/migrations",
-          ),
-        ).rejects.toThrow(
-          "Invalid destination path: must be within userData directory",
-        );
-
-        expect(mockLogger.error).toHaveBeenCalledWith(
-          "Destination path outside userData directory",
-        );
-      });
-
-      it("should throw error if no SQL files found in source", async () => {
-        mockFileSystemBridge.getDirectoryStats.mockResolvedValue({
-          exists: true,
-          isDirectory: true,
-        });
-        mockFileSystemBridge.readdir.mockResolvedValue([
-          "readme.txt",
-          "config.json",
-        ]);
-
-        await expect(
-          (testServices as any).validateMigrationPaths(
-            "/source/migrations",
-            "/mock/userdata/migrations",
-          ),
-        ).rejects.toThrow(
-          "No migration files (.sql) found in source directory",
-        );
-
-        expect(mockLogger.warn).toHaveBeenCalledWith(
-          "No .sql files found in source migrations path",
-          expect.objectContaining({
-            sourcePath: "/source/migrations",
-            filesFound: 2,
-          }),
-        );
-      });
-
-      it("should handle file system errors gracefully", async () => {
-        const fsError = new Error("File system access denied");
-        mockFileSystemBridge.getDirectoryStats.mockRejectedValue(fsError);
-
-        await expect(
-          (testServices as any).validateMigrationPaths(
-            "/source/migrations",
-            "/mock/userdata/migrations",
-          ),
-        ).rejects.toThrow("File system access denied");
-
-        expect(mockLogger.error).toHaveBeenCalledWith(
-          "Migration path validation failed",
-          fsError,
-        );
-      });
-    });
-  });
-
-  describe("migration copying functionality", () => {
-    describe("copyMigrationsToUserData", () => {
-      let testServices: MainProcessServices;
-      let mockFileSystemBridge: any;
-      let mockLogger: any;
-
-      beforeEach(() => {
-        testServices = new MainProcessServices();
-        mockLogger = {
-          debug: jest.fn(),
-          info: jest.fn(),
-          warn: jest.fn(),
-          error: jest.fn(),
-        };
-        mockFileSystemBridge = {
-          getDirectoryStats: jest.fn(),
-          readdir: jest.fn(),
-          readFile: jest.fn(),
-          writeFile: jest.fn(),
-          ensureDirectoryExists: jest.fn(),
-          copyFile: jest.fn(),
-          exists: jest.fn(),
-          ensureDir: jest.fn(),
-        };
-        (testServices as any).logger = mockLogger;
-        (testServices as any).fileSystemBridge = mockFileSystemBridge;
-      });
-
-      it("should successfully copy migration files from source to destination", async () => {
-        mockFileSystemBridge.getDirectoryStats.mockResolvedValue({
-          exists: true,
-          isDirectory: true,
-        });
-        mockFileSystemBridge.readdir.mockResolvedValue([
-          "001_initial.sql",
-          "002_add_users.sql",
-          "003_add_indexes.sql",
-          "readme.txt", // Should be filtered out
-        ]);
-        mockFileSystemBridge.readFile
-          .mockResolvedValueOnce("CREATE TABLE conversations;")
-          .mockResolvedValueOnce("CREATE TABLE users;")
-          .mockResolvedValueOnce("CREATE INDEX idx_users_email;");
-
-        await (testServices as any).copyMigrationsToUserData();
-
-        expect(mockFileSystemBridge.ensureDirectoryExists).toHaveBeenCalledWith(
-          expect.stringContaining("migrations"),
-        );
-        expect(mockFileSystemBridge.writeFile).toHaveBeenCalledTimes(3);
-        expect(mockFileSystemBridge.writeFile).toHaveBeenCalledWith(
-          expect.stringContaining("001_initial.sql"),
-          "CREATE TABLE conversations;",
-        );
-        expect(mockFileSystemBridge.writeFile).toHaveBeenCalledWith(
-          expect.stringContaining("002_add_users.sql"),
-          "CREATE TABLE users;",
-        );
-        expect(mockLogger.info).toHaveBeenCalledWith(
-          "Migration files copied successfully",
-          expect.objectContaining({
-            fileCount: 3,
-            durationMs: expect.any(Number),
-          }),
-        );
-      });
-
-      it("should handle missing source directory gracefully", async () => {
-        mockFileSystemBridge.getDirectoryStats.mockResolvedValue({
-          exists: false,
-          isDirectory: false,
-        });
-
-        await (testServices as any).copyMigrationsToUserData();
-
-        expect(mockLogger.warn).toHaveBeenCalledWith(
-          "Source migrations directory not found",
-          expect.objectContaining({
-            sourcePath: expect.any(String),
-          }),
-        );
-        expect(mockFileSystemBridge.writeFile).not.toHaveBeenCalled();
-      });
-
-      it("should handle empty source directory gracefully", async () => {
-        mockFileSystemBridge.getDirectoryStats.mockResolvedValue({
-          exists: true,
-          isDirectory: true,
-        });
-        mockFileSystemBridge.readdir.mockResolvedValue([]);
-
-        await (testServices as any).copyMigrationsToUserData();
-
-        expect(mockLogger.warn).toHaveBeenCalledWith(
-          "No migration files found in source directory",
-          expect.objectContaining({
-            sourcePath: expect.any(String),
-          }),
-        );
-        expect(mockFileSystemBridge.writeFile).not.toHaveBeenCalled();
-      });
-
-      it("should filter files by migration pattern", async () => {
-        mockFileSystemBridge.getDirectoryStats.mockResolvedValue({
-          exists: true,
-          isDirectory: true,
-        });
-        mockFileSystemBridge.readdir.mockResolvedValue([
-          "001_initial.sql", // Should be copied
-          "002_add_users.sql", // Should be copied
-          "1_invalid.sql", // Should be filtered out (doesn't match pattern)
-          "invalid_migration.sql", // Should be filtered out
-          "readme.txt", // Should be filtered out
-          "config.json", // Should be filtered out
-        ]);
-        mockFileSystemBridge.readFile
-          .mockResolvedValueOnce("CREATE TABLE conversations;")
-          .mockResolvedValueOnce("CREATE TABLE users;");
-
-        await (testServices as any).copyMigrationsToUserData();
-
-        expect(mockFileSystemBridge.writeFile).toHaveBeenCalledTimes(2);
-        expect(mockLogger.info).toHaveBeenCalledWith(
-          "Migration files copied successfully",
-          expect.objectContaining({
-            fileCount: 2,
-          }),
-        );
-      });
-
-      it("should handle file system errors during copying", async () => {
-        mockFileSystemBridge.getDirectoryStats.mockResolvedValue({
-          exists: true,
-          isDirectory: true,
-        });
-        mockFileSystemBridge.readdir.mockResolvedValue(["001_initial.sql"]);
-        const fsError = new Error("Permission denied");
-        mockFileSystemBridge.readFile.mockRejectedValue(fsError);
-
-        await expect(
-          (testServices as any).copyMigrationsToUserData(),
-        ).rejects.toThrow("Migration file copying failed: Permission denied");
-
-        expect(mockLogger.error).toHaveBeenCalledWith(
-          "Failed to copy migration files",
-          fsError,
-        );
-      });
-
-      it("should track timing performance", async () => {
-        mockFileSystemBridge.getDirectoryStats.mockResolvedValue({
-          exists: true,
-          isDirectory: true,
-        });
-        mockFileSystemBridge.readdir.mockResolvedValue(["001_initial.sql"]);
-        mockFileSystemBridge.readFile.mockResolvedValue("CREATE TABLE users;");
-
-        await (testServices as any).copyMigrationsToUserData();
-
-        expect(mockLogger.info).toHaveBeenCalledWith(
-          "Migration files copied successfully",
-          expect.objectContaining({
-            durationMs: expect.any(Number),
-          }),
-        );
-      });
-    });
-
-    describe("runDatabaseMigrations with file copying", () => {
-      let mockMigrationService: any;
-      let mockFileSystemBridge: any;
-      let mockLogger: any;
-
-      beforeEach(() => {
-        services = new MainProcessServices();
-        mockMigrationService = services.migrationService;
-        mockLogger = {
-          debug: jest.fn(),
-          info: jest.fn(),
-          warn: jest.fn(),
-          error: jest.fn(),
-        };
-        mockFileSystemBridge = {
-          getDirectoryStats: jest.fn(),
-          readdir: jest.fn(),
-          readFile: jest.fn(),
-          writeFile: jest.fn(),
-          ensureDirectoryExists: jest.fn(),
-          copyFile: jest.fn(),
-          exists: jest.fn(),
-          ensureDir: jest.fn(),
-        };
-        (services as any).logger = mockLogger;
-        (services as any).fileSystemBridge = mockFileSystemBridge;
-      });
-
-      it("should ensure migrations are copied before running migrations", async () => {
-        const mockResult = {
-          success: true,
-          migrationsRun: 2,
-          currentVersion: 2,
-        };
-        mockMigrationService.runMigrations.mockResolvedValue(mockResult);
-
-        // Mock successful copy scenario
-        mockFileSystemBridge.getDirectoryStats.mockResolvedValue({
-          exists: true,
-          isDirectory: true,
-        });
-        mockFileSystemBridge.readdir.mockResolvedValue(["001_initial.sql"]);
-        mockFileSystemBridge.readFile.mockResolvedValue("CREATE TABLE users;");
-
-        await services.runDatabaseMigrations();
-
-        expect(mockFileSystemBridge.getDirectoryStats).toHaveBeenCalled();
-        expect(mockFileSystemBridge.ensureDirectoryExists).toHaveBeenCalled();
-        expect(mockFileSystemBridge.writeFile).toHaveBeenCalled();
-        expect(mockMigrationService.runMigrations).toHaveBeenCalled();
-        expect(mockLogger.info).toHaveBeenCalledWith(
-          "Migration files copied successfully",
-          expect.any(Object),
-        );
-      });
-
-      it("should handle copying errors during migration run", async () => {
-        const copyError = new Error("File system error");
-        mockFileSystemBridge.getDirectoryStats.mockRejectedValue(copyError);
-
-        await expect(services.runDatabaseMigrations()).rejects.toThrow(
-          "Migration execution failed: Migration file copying failed: File system error",
-        );
-
-        expect(mockMigrationService.runMigrations).not.toHaveBeenCalled();
-      });
+      expect(mockRunDatabaseMigrations).toHaveBeenCalled();
     });
   });
 });
